requirements.txt
flask==3.0.2
flask-wtf==1.2.2
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0
wtforms==3.2.1
